// temp/frontend/ts/proto/jon/jon_shared_data.ts
export interface JonGUIState {
  protocolVersion: number;
  system: JonGuiDataSystem | undefined;
  meteoInternal: JonGuiDataMeteo | undefined;
  meteoExternal: JonGuiDataMeteo | undefined;
  lrf: JonGuiDataLrf | undefined;
  time: JonGuiDataTime | undefined;
  gps: JonGuiDataGps | undefined;
  compass: JonGuiDataCompass | undefined;
  rotary: JonGuiDataRotary | undefined;
  cameraDay: JonGuiDataCameraDay | undefined;
  cameraHeat: JonGuiDataCameraHeat | undefined;
  compassCalibration: JonGuiDataCompassCalibration | undefined;
  environment: JonGuiDataEnvironment | undefined;
  power: JonGuiDataPower | undefined;
}