// temp/frontend/ts/proto/jon/jon_shared_data_power.ts
export interface JonGuiDataPowerModuleState {
  voltage: number;
  current: number;
  power: number;
  isAlarm: boolean;
  canCmdAddress: number;
  canDataAddress: number;
  isPowerOn: boolean;
  canDevice: JonGuiDataPowerCanDevice;
}

export interface JonGuiDataPower {
  s0: JonGuiDataPowerModuleState | undefined;
  s1: JonGuiDataPowerModuleState | undefined;
  s2: JonGuiDataPowerModuleState | undefined;
  s3: JonGuiDataPowerModuleState | undefined;
  s4: JonGuiDataPowerModuleState | undefined;
  s5: JonGuiDataPowerModuleState | undefined;
  s6: JonGuiDataPowerModuleState | undefined;
  s7: JonGuiDataPowerModuleState | undefined;
  meteo: JonGuiDataMeteo | undefined;
}

function createBaseJonGuiDataPowerModuleState(): JonGuiDataPowerModuleState {
  return {
    voltage: 0,
    current: 0,
    power: 0,
    isAlarm: false,
    canCmdAddress: 0,
    canDataAddress: 0,
    isPowerOn: false,
    canDevice: 0,
  };
}

export const JonGuiDataPowerModuleState = {
  encode(message: JonGuiDataPowerModuleState, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.voltage !== 0) {
      writer.uint32(8).int32(message.voltage);
    }
    if (message.current !== 0) {
      writer.uint32(16).int32(message.current);
    }
    if (message.power !== 0) {
      writer.uint32(24).int32(message.power);
    }
    if (message.isAlarm !== false) {
      writer.uint32(32).bool(message.isAlarm);
    }
    if (message.canCmdAddress !== 0) {
      writer.uint32(40).int32(message.canCmdAddress);
    }
    if (message.canDataAddress !== 0) {
      writer.uint32(48).int32(message.canDataAddress);
    }
    if (message.isPowerOn !== false) {
      writer.uint32(56).bool(message.isPowerOn);
    }
    if (message.canDevice !== 0) {
      writer.uint32(64).int32(message.canDevice);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): JonGuiDataPowerModuleState {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseJonGuiDataPowerModuleState();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break;
          }

          message.voltage = reader.int32();
          continue;
        case 2:
          if (tag !== 16) {
            break;
          }

          message.current = reader.int32();
          continue;
        case 3:
          if (tag !== 24) {
            break;
          }

          message.power = reader.int32();
          continue;
        case 4:
          if (tag !== 32) {
            break;
          }

          message.isAlarm = reader.bool();
          continue;
        case 5:
          if (tag !== 40) {
            break;
          }

          message.canCmdAddress = reader.int32();
          continue;
        case 6:
          if (tag !== 48) {
            break;
          }

          message.canDataAddress = reader.int32();
          continue;
        case 7:
          if (tag !== 56) {
            break;
          }

          message.isPowerOn = reader.bool();
          continue;
        case 8:
          if (tag !== 64) {
            break;
          }

          message.canDevice = reader.int32() as any;
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): JonGuiDataPowerModuleState {
    return {
      voltage: isSet(object.voltage) ? globalThis.Number(object.voltage) : 0,
      current: isSet(object.current) ? globalThis.Number(object.current) : 0,
      power: isSet(object.power) ? globalThis.Number(object.power) : 0,
      isAlarm: isSet(object.isAlarm) ? globalThis.Boolean(object.isAlarm) : false,
      canCmdAddress: isSet(object.canCmdAddress) ? globalThis.Number(object.canCmdAddress) : 0,
      canDataAddress: isSet(object.canDataAddress) ? globalThis.Number(object.canDataAddress) : 0,
      isPowerOn: isSet(object.isPowerOn) ? globalThis.Boolean(object.isPowerOn) : false,
      canDevice: isSet(object.canDevice) ? jonGuiDataPowerCanDeviceFromJSON(object.canDevice) : 0,
    };
  },

  toJSON(message: JonGuiDataPowerModuleState): unknown {
    const obj: any = {};
    if (message.voltage !== 0) {
      obj.voltage = Math.round(message.voltage);
    }
    if (message.current !== 0) {
      obj.current = Math.round(message.current);
    }
    if (message.power !== 0) {
      obj.power = Math.round(message.power);
    }
    if (message.isAlarm !== false) {
      obj.isAlarm = message.isAlarm;
    }
    if (message.canCmdAddress !== 0) {
      obj.canCmdAddress = Math.round(message.canCmdAddress);
    }
    if (message.canDataAddress !== 0) {
      obj.canDataAddress = Math.round(message.canDataAddress);
    }
    if (message.isPowerOn !== false) {
      obj.isPowerOn = message.isPowerOn;
    }
    if (message.canDevice !== 0) {
      obj.canDevice = jonGuiDataPowerCanDeviceToJSON(message.canDevice);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<JonGuiDataPowerModuleState>, I>>(base?: I): JonGuiDataPowerModuleState {
    return JonGuiDataPowerModuleState.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<JonGuiDataPowerModuleState>, I>>(object: I): JonGuiDataPowerModuleState {
    const message = createBaseJonGuiDataPowerModuleState();
    message.voltage = object.voltage ?? 0;
    message.current = object.current ?? 0;
    message.power = object.power ?? 0;
    message.isAlarm = object.isAlarm ?? false;
    message.canCmdAddress = object.canCmdAddress ?? 0;
    message.canDataAddress = object.canDataAddress ?? 0;
    message.isPowerOn = object.isPowerOn ?? false;
    message.canDevice = object.canDevice ?? 0;
    return message;
  },
};
// ...
type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };